refactor diffViews
